Deception/project/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
pytest==8.4.1
httpx==0.28.1
ff3==1.0.3
